removed cocoapods
