+++ VaccineCard/VaccineCard/Services/JWTService.swift
@@ -1,0 +1,13 @@
+//
+//  JWTService.swift
+//  VaccineCard
+//
+//  Created by Amir Shayegh on 2021-09-02.
+//
+
+import Foundation
+import SwiftJWT
+
+class JWTService {
+    
+}
